Fix restat dirty check logic
In a normal dependency scan (see DependencyScan::RecomputeDirty) we
mark all outputs of an Edge as dirty if any of the outputs is dirty.
This is the correct behavior because if any output is dirty, we will
run the command for that Edge and that can touch any of the outputs of
the Edge and so all the outputs should be marked dirty.

When updating the dirty state of Node's for a restat check, we were not
applying this logic, instead only those outputs that were actually
"dirty" were marked dirty.